feat: use a factory to instantiate classes with common required context
