+++ typescript/cli/src/rebalancer/metrics/utils.ts
@@ -1,3 +1,4 @@
+import { Token } from '@hyperlane-xyz/sdk';
 import { rootLogger } from '@hyperlane-xyz/utils';
 
 export const logger = rootLogger.child({ module: 'warp-balance-monitor' });
@@ -8,4 +9,8 @@ export async function tryFn(fn: () => Promise<void>, context: string) {
   } catch (e) {
     logger.error(`Error in ${context}`, e);
   }
+}
+
+export function formatBigInt(warpToken: Token, num: bigint): number {
+  return warpToken.amount(num).getDecimalFormattedAmount();
 }
